\name{Astrand}

\alias{Astrand}

\title{Physical working capacity in relation to sex and age}

\description{
In an interesting early study on changes in 'working capacity' with sex and age, Astrand (1952) recorded measurements of a variety of physiological parameters in subjects who undertook exercise at maximal intensity.
}

\references{
Astrand, Per-Olof. Experimental studies of physical working capacity in relation to sex and age. Munksgaard Forlag.
}

\examples{
\dontrun{
plot(Heart_rate ~ Age, data = Astrand)
}}

\keyword{iplot}
\keyword{dynamic}
