\name{doctors_smoking}

\alias{doctors_smoking}

\title{Smoking habits among doctors}

\description{
These data are from a famous study, initiated by Doll & Hill, which surveyed UK doctors and related their smoking habits to subsequent coronary-related deaths. The data in a form reduced to person-years exposure was reported by Breslow (1985).
}

\references{
Breslow, N.E. (1985) Cohort Analysis in Epidemiology. In 'A Celebration of Statistics', 
A.C. Atkinson and S.E. Fienberg (editors), 109-143. Springer-Verlag.

Doll, R. and Hill, A.B. (1966) Mortality of British doctors in relation to smoking: Observations on coronary thrombosis. National Cancer Institute Monograph, 19, 205-268.
}

\keyword{iplot}
\keyword{dynamic}
