\name{gulls.all}

\alias{gulls.all}

\title{Data used by rp.gulls}

\description{
  These data are used internally by rp.gulls.
}

\references{rpanel: Simple interactive controls for R functions using
the tcltk package
(http://www.stats.gla.ac.uk/~adrian/rpanel/)}

\examples{
  rp.gulls()
}

\keyword{iplot}
\keyword{dynamic}

