\name{Scottish_referendum}

\alias{Scottish_referendum}

\title{The Scottish referendum on independence}

\description{
These data record the results of the Scottish referendum on independence held in 2014. The variables include the votes counted in each local council area along with other social characteristics.
}

\references{
Scottish Independence Referendum 2014: Analysis of results.
Research Paper 14/50, 30 September 2014.
https://commonslibrary.parliament.uk/research-briefings/rp14-50/}

\keyword{iplot}
\keyword{dynamic}

